refactor(skills): split model-visible tools into boundary owners (S3)
Split the 552-line src/core/skills/tools.ts into seven modules under
src/core/skills/tools/: definitions (tool JSON schemas), arguments
(strict model-tool argument parsing), activated-skill (shared contract:
runtime options, ValidSkill, activation gate, path resolution, block
formatter), activation (activate_skill executor — no process import),
resource (read_skill_resource executor — no interpreter/process import),
script (run_skill_script executor through Process Runtime), and index
(public barrel). Consumer imports from ./tools automatically resolve to
./tools/index.ts, so no consumer paths change.
