+++ src/core/skills/tools/index.ts
@@ -1,0 +1,9 @@
+// Public tool facade: re-exports the three tool definitions, executors, and
+// shared types. Consumers import from "./tools" (this barrel).
+
+export { createActivateSkillToolDefinition, readSkillResourceToolDefinition, runSkillScriptToolDefinition } from "./definitions";
+export { executeActivateSkillTool } from "./activation";
+export { executeReadSkillResourceTool } from "./resource";
+export { executeRunSkillScriptTool } from "./script";
+export { formatSkillActivationBlock } from "./activated-skill";
+export type { SkillToolRuntimeOptions, ValidSkill } from "./activated-skill";
